+++ lib/src/_index.dart
@@ -2,6 +2,9 @@
 // End user is granted a non-exclusive non-transferable license the ("License")
 // to use GM Consult's proprietary software (the "Software").
 
-export 'text_analyzer_configuration/_index.dart';
-export 'text_analyzer/_index.dart';
+export 'text_tokenizer.dart';
+export 'text_document.dart';
+export 'token.dart';
 export 'typedefs.dart';
+export 'text_analyzer.dart';
+export 'english.dart';