+++ lib/src/text_analyzer/_index.dart
@@ -1,7 +1,0 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
-// End user is granted a non-exclusive non-transferable license the ("License")
-// to use GM Consult's proprietary software (the "Software").
-
-export 'text_analyzer.dart';
-export 'text_document.dart';
-export 'token.dart';