+++ lib/src/text_analyzer_configuration/_index.dart
@@ -1,6 +1,0 @@
-// Copyright ©2022, GM Consult (Pty) Ltd. All rights reserved
-// End user is granted a non-exclusive non-transferable license the ("License")
-// to use GM Consult's proprietary software (the "Software").
-
-export 'text_analyzer_configuration.dart';
-export 'english.dart';